backend/requirements.txt
fastapi
uvicorn
pyjwt
passlib[bcrypt]
python-multipart
sqlalchemy
pymysql
alembic
crpytography